update pipeline controller
