+++ jenkins/build_drum.groovy
@@ -1,22 +1,14 @@
 node('multi-executor && ubuntu:focal'){
-    echo "Build DRUM"
-    stage('Checkout') {
-        cleanWs()
+    stage('build_drum') {
         checkout scm
-    }
-
-    stage('Build DRUM') {
-        echo "I'm build stage"
         sh """
-            echo "I'm build stage from shell"
             echo $WORKSPACE
-
             source "tools/image-build-utils.sh"
             build_drum
             ls -la custom_model_runner/dist/*
         """
-    }
-    dir ('custom_model_runner/dist') {
-        stash includes: '*', name: 'drum_wheel'
+        dir ('custom_model_runner/dist') {
+            stash includes: '*', name: 'drum_wheel'
+        }
     }
 }
