+++ pipelineController.yaml
@@ -15,6 +15,13 @@ test_integration_all_in_one: &test_integration_all_in_one
     PIPELINE_CONTROLLER: 1
   phase: 2
 
+test_integration_all_in_one_bare_metal: &test_integration_all_in_one_bare_metal
+  taskName: test_integration_all_in_one_bare_metal
+  definition: jenkins/run_integration_tests_without_container.groovy
+  environment:
+    PIPELINE_CONTROLLER: 1
+  phase: 2
+
 test_inference_model_templates: &test_inference_model_templates
   taskName: test_inference_model_templates
   definition: jenkins/test_inference_model_templates.groovy
@@ -42,16 +49,35 @@ repositoryTasks:
       - regex: ".*black_check.*"
         script: *black_check
       - regex: ".*build_drum.*"
-        script: *build_drum
+        script:
+          - *black_check
+          - *build_drum
+      - regex: ".*test_integration_all_in_one.*"
+        script:
+          - *build_drum
+          - *test_integration_all_in_one
+      - regex: ".*test_integration_all_in_one_bare_metal.*"
+        script:
+          - *build_drum
+          - *test_integration_all_in_one_bare_metal
       - regex: ".*test_inference_model_templates.*"
         script:
           - *build_drum
           - *test_inference_model_templates
+      - regex: ".*test_training_model_templates.*"
+        script:
+          - *build_drum
+          - *test_training_model_templates
+      - regex: ".*test_drop_in_envs.*"
+        script:
+          - *build_drum
+          - *test_drop_in_envs
     change:
       - changedFilesRegex: '.*'
         script:
           - *black_check
           - *build_drum
+          - *test_integration_all_in_one_bare_metal
           - *test_integration_all_in_one
           - *test_inference_model_templates
           - *test_training_model_templates
